Refactor IL2CPP class, field, and method structures
Introduces and implements Clone and Debug traits for Class, Field, and Method structs. Refactors Field and Method to include constructors and additional fields. Updates class_get_fields API to use *mut u8 for iterators, and modifies Cache to parse and store fields for each class. These changes improve type safety, debuggability, and prepare for more robust IL2CPP metadata handling.

+++ src/il2cpp/il2cpp_sys/fn_types.rs
@@ -23,7 +23,7 @@ pub type Il2CppImageGetClassCountFn = unsafe extern "C" fn(image: Il2CppImage) -
 pub type Il2CppImageGetClassFn =
     unsafe extern "C" fn(image: Il2CppImage, index: usize) -> Il2CppClass;
 pub type Il2CppClassGetFieldsFn =
-    unsafe extern "C" fn(klass: Il2CppClass, iter: *mut usize) -> *mut u8; // Il2CppField
+    unsafe extern "C" fn(klass: Il2CppClass, iter: *mut u8) -> *mut u8; // Il2CppField
 pub type Il2CppFieldGetNameFn = unsafe extern "C" fn(field: *mut u8) -> *const i8;
 pub type Il2CppFieldGetOffsetFn = unsafe extern "C" fn(field: *mut u8) -> u32;
 pub type Il2CppMethodGetParamCountFn = unsafe extern "C" fn(method: Il2CppMethodInfo) -> u32;
